Move the padded-leaf claim reduction to the prover
`unpad_leaf_claim` turns a claim on a padded witness back into one on the
witness itself. Only the prover holds such claims: both copies — the product
check's and the fractional-addition check's — are called from `binius-ip-prover`
alone, and neither is reachable from any verifier.

Move them to the crate that uses them. The forward map `pad_leaf_fraction`
stays: a verifier rebuilding a padded batch's leaf claim from transparent parts
does apply it.

This also takes the two `invert_or_zero` calls out of the verifier crates. They
were the only ones there, and each divides by a padding weight the line above
asserts is non-zero.
